+++ src/e2e/github-actions-firefox.conf.js
@@ -1,0 +1,10 @@
+const config = require("./protractor.conf").config;
+
+config.capabilities = {
+  browserName: "firefox",
+  "moz:firefoxOptions": {
+    args: ["-headless"]
+  }
+};
+
+exports.config = config;