+++ src/e2e/src/integration/upload-toolbar.spec.ts
@@ -1,5 +1,5 @@
 import { simulateDrop } from "../utils/drag-event";
-import { browser, logging } from "protractor";
+import { browser } from "protractor";
 import { spawn, ChildProcess } from "child_process";
 import { Dashboard } from "../support/dashboard.po";
 import { UploadToolbarPO } from "../support/ngx-fileupload-ui/upload-toolbar";
@@ -107,18 +107,6 @@ describe("Ngx Fileupload Upload Toolbar", () => {
 
         /** start uploads */
         await uploadToolbar.removeAll();
-    });
-
-    afterEach(async () => {
-        // Assert that there are no errors emitted from the browser
-        let logs = await browser.manage().logs().get(logging.Type.BROWSER);
-
-        /** we expect an error in browser with state of 401 so we filter this out */
-        const error = "responded with a status of 401 (Unauthorized)";
-        logs = logs.filter((entry: logging.Entry) => !entry.message.endsWith(error));
-
-        expect(logs)
-            .not.toContain(jasmine.objectContaining({ level: logging.Level.SEVERE } as logging.Entry));
     });
 
     afterAll(() => {